Fix `TimeDeltaMissingUnitWarning` in `plot_airmass`
The function `plot_airmass` triggers the warning `TimeDeltaMissingUnitWarning`. This issue arises from unit checks implemented in newer versions of Astropy (refer to commit d8968b29a209fdaebd0f48ff30863718bca9db78).

The warning is generated because a time value without a unit is being added to an Astropy Time object. To resolve this warning, it is recommended to add the unit `u.hour` to the time value.